JDBC access to arrow data
Add an new interface `ArrowResultSet` to be optionally implemented by
JDBC drivers to allow `java.sql.ResultSet` to be unwrapped into a
sequence of Arrow vectors and bypass JDBC serialization.
