+++ java/adapter/jdbc/src/main/java/org/apache/arrow/adapter/jdbc/ArrowVectorIterator.java
@@ -18,164 +18,27 @@
 package org.apache.arrow.adapter.jdbc;
 
 import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.util.Iterator;
 
-import org.apache.arrow.adapter.jdbc.consumer.CompositeJdbcConsumer;
-import org.apache.arrow.adapter.jdbc.consumer.JdbcConsumer;
-import org.apache.arrow.util.Preconditions;
+import org.apache.arrow.sql.ArrowResultSet;
 import org.apache.arrow.vector.VectorSchemaRoot;
-import org.apache.arrow.vector.types.pojo.Schema;
-import org.apache.arrow.vector.util.ValueVectorUtility;
 
 /**
- * VectorSchemaRoot iterator for partially converting JDBC data.
+ * VectorSchemaRoot iterator over JDBC data.
  */
-public class ArrowVectorIterator implements Iterator<VectorSchemaRoot>, AutoCloseable {
-
-  private final ResultSet resultSet;
-  private final JdbcToArrowConfig config;
-
-  private final Schema schema;
-  private final ResultSetMetaData rsmd;
-
-  private final JdbcConsumer[] consumers;
-  final CompositeJdbcConsumer compositeConsumer;
-
-  private VectorSchemaRoot nextBatch;
-
-  private final int targetBatchSize;
-
-  /**
-   * Construct an instance.
-   */
-  private ArrowVectorIterator(ResultSet resultSet, JdbcToArrowConfig config) throws SQLException {
-    this.resultSet = resultSet;
-    this.config = config;
-    this.schema = JdbcToArrowUtils.jdbcToArrowSchema(resultSet.getMetaData(), config);
-    this.targetBatchSize = config.getTargetBatchSize();
-
-    rsmd = resultSet.getMetaData();
-    consumers = new JdbcConsumer[rsmd.getColumnCount()];
-    this.compositeConsumer = new CompositeJdbcConsumer(consumers);
-  }
-
-  private void initialize() throws SQLException {
-    // create consumers
-    for (int i = 1; i <= consumers.length; i++) {
-      consumers[i - 1] = JdbcToArrowUtils.getConsumer(resultSet, i, resultSet.getMetaData().getColumnType(i),
-          null, config);
-    }
-
-    load(createVectorSchemaRoot());
-  }
+public interface ArrowVectorIterator extends Iterator<VectorSchemaRoot>, AutoCloseable {
+  @Override
+  void close();
 
   /**
    * Create a ArrowVectorIterator to partially convert data.
    */
-  public static ArrowVectorIterator create(
-      ResultSet resultSet,
-      JdbcToArrowConfig config)
+  static ArrowVectorIterator create(ResultSet resultSet, JdbcToArrowConfig config)
       throws SQLException {
-
-    ArrowVectorIterator iterator = new ArrowVectorIterator(resultSet, config);
-    try {
-      iterator.initialize();
-      return iterator;
-    } catch (Exception e) {
-      iterator.close();
-      throw new RuntimeException("Error occurred while creating iterator.", e);
+    if (resultSet.isWrapperFor(ArrowResultSet.class)) {
+      return DirectArrowVectorIterator.create(resultSet, config);
     }
-  }
-
-  private void consumeData(VectorSchemaRoot root) {
-    // consume data
-    try {
-      int readRowCount = 0;
-      if (targetBatchSize == JdbcToArrowConfig.NO_LIMIT_BATCH_SIZE) {
-        while (resultSet.next()) {
-          ValueVectorUtility.ensureCapacity(root, readRowCount + 1);
-          compositeConsumer.consume(resultSet);
-          readRowCount++;
-        }
-      } else {
-        while (readRowCount < targetBatchSize && resultSet.next()) {
-          compositeConsumer.consume(resultSet);
-          readRowCount++;
-        }
-      }
-
-
-      root.setRowCount(readRowCount);
-    } catch (Exception e) {
-      compositeConsumer.close();
-      throw new RuntimeException("Error occurred while consuming data.", e);
-    }
-  }
-
-  private VectorSchemaRoot createVectorSchemaRoot() {
-    VectorSchemaRoot root = null;
-    try {
-      root = VectorSchemaRoot.create(schema, config.getAllocator());
-      if (config.getTargetBatchSize() != JdbcToArrowConfig.NO_LIMIT_BATCH_SIZE) {
-        ValueVectorUtility.preAllocate(root, config.getTargetBatchSize());
-      }
-    } catch (Exception e) {
-      if (root != null) {
-        root.close();
-      }
-      throw new RuntimeException("Error occurred while creating schema root.", e);
-    }
-    return root;
-  }
-
-  // Loads the next schema root or null if no more rows are available.
-  private void load(VectorSchemaRoot root) throws SQLException {
-
-    for (int i = 1; i <= consumers.length; i++) {
-      consumers[i - 1].resetValueVector(root.getVector(rsmd.getColumnName(i)));
-    }
-
-    consumeData(root);
-
-    if (root.getRowCount() == 0) {
-      root.close();
-      nextBatch = null;
-    } else {
-      nextBatch = root;
-    }
-  }
-
-  @Override
-  public boolean hasNext() {
-    return nextBatch != null;
-  }
-
-  /**
-   * Gets the next vector. The user is responsible for freeing its resources.
-   */
-  @Override
-  public VectorSchemaRoot next() {
-    Preconditions.checkArgument(hasNext());
-    VectorSchemaRoot returned = nextBatch;
-    try {
-      load(createVectorSchemaRoot());
-    } catch (Exception e) {
-      close();
-      throw new RuntimeException("Error occurred while getting next schema root.", e);
-    }
-    return returned;
-  }
-
-  /**
-   * Clean up resources.
-   */
-  @Override
-  public void close() {
-    if (nextBatch != null) {
-      nextBatch.close();
-    }
-    compositeConsumer.close();
+    return IndirectArrowVectorIterator.create(resultSet, config);
   }
 }
